lambda_functions/stac/requirements.txt
boto3==1.6.23
pyproj==1.9.5.1
parse==1.9.0
python_dateutil==2.7.5
PyYAML==3.13
